initial, minimal schema
Generated via mysqldump w/ --compact, piped to grep -v '/*!40101 SET '

Very basic structure: users, projects, and skills. A skill can be
assigned to a user, and to a project. It may make sense to have an extra
column in project_skill to determine if that skill is being met or not,
so we can tie that to calls for action.

User table can be filled in as we log in via Persona. No space for
access control, currently, preferably to be built out via roles to the
degree we want them.

Not currently sure on the best path for Open Badges, though in theory
that structure can be incorporated here as well.